+++ src/index.ts
@@ -23,3 +23,4 @@ export * from "./components/PreviewProfile";
 export * from "./components/SmallProfileImage";
 export * from "./components/FluctuationComponent";
 export * from "./components/CollateralSection";
+export * from "./components/Tooltip";