+++ src/stories/CardComponent.stories.tsx
@@ -2,19 +2,24 @@ import React from "react";
 import { Story } from "@storybook/react";
 
 import { CardComponent, ICardComponent } from "../components/CardComponent";
+import { Tooltip } from "../components/Tooltip";
+import InfoIcon from "../../public/assets/info-icon.svg";
 
 export default {
   title: "Card Component",
   component: CardComponent,
 };
 
+const exampleTooltip = <Tooltip icon={InfoIcon} text="This is a tooltip" />;
+
 const Template: Story<ICardComponent> = (args) => <CardComponent {...args} />;
 
 export const CardComponentItem = Template.bind({});
 CardComponentItem.args = {
-  color: '#26A66B',
-  label: 'Total',
-  value: '$2,000',
-  fluctuation: '+3.5',
-  fluctuationValue: '+$2,400'
+  color: "#26A66B",
+  label: "Total",
+  value: "$2,000",
+  fluctuation: "+3.5",
+  fluctuationValue: "+$2,400",
+  tooltip: exampleTooltip,
 };